Add basic hit test and fix render bug
There was a problem where all my cards where getting the same Z
coordinate in the final render. The reason for this was that in the card
model, I was scaling the X and Y axis, but I set the Z axis scaling to
0. It should have been set to 1, and doing so made it so I had to fix
some other issues around the rendering, but now it should be working.